permissions-policy/payment-disabled-by-permissions-policy.https.sub.html

<!DOCTYPE html>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/permissions-policy/resources/permissions-policy.js"></script>
    <script>
        "use strict";
        const same_origin_src =
            "/permissions-policy/resources/permissions-policy-payment.html";
        const cross_origin_src =
            "https://{{hosts[alt][]}}:{{ports[https][0]}}" + same_origin_src;
        const header = 'permissions policy header "payment=()"';

        test(() => {
            const supportedInstruments = [{ supportedMethods: "visa" }];
            const details = {
                total: {
                    label: "Test",
                    amount: { currency: "USD", value: "5.00" },
                },
            };
            assert_throws_dom("SecurityError", () => {
                new PaymentRequest(supportedInstruments, details);
            });
        }, header + " disallows the top-level document.");

        promise_test((t) => {
            return test_feature_availability(
                "PaymentRequest()",
                t,
                same_origin_src,
                expect_feature_unavailable_default,
                undefined,
                undefined,
                true
            );
        }, header + " disallows same-origin iframes.");

        promise_test((t) => {
            return test_feature_availability(
                "PaymentRequest()",
                t,
                cross_origin_src,
                expect_feature_unavailable_default,
                undefined,
                undefined,
                true
            );
        }, header + " disallows cross-origin iframes.");
    </script>
</body>
